Edit parts of fares impute calc
Made small edits to a drop function call to ensure the column names are searched for in the column names list and not the numeric index. Also investigated the reason for the test_fares_imputation test failing; ~3% of the final results from fares imputation are different due to the changeover to Python 3. This is because Python 3 uses a different rounding method to SAS which causes some values to be rounded up instead of down and vice versa. Although this test fails, the Python 3 method is the more widely accepted version of rounding so is more correct.