Rewrite non_response_weight and add test
Calculate_non_response_weight has been rewritten so that the code will
execute successfully and produce the correct calculated values. This has
been confirmed by the addition of a unit test to ensure the output
matches the SAS output. Note that the output columns are in a different
order and the test used check_like as TRUE to avoid this issue.


Former-commit-id: 919804276aa2a3a3cbaa5ac784f67fe80af85393