Fix missing rows error in minimums weight
Found that missing vlaues in the SAS data for MINS_CTRY_GRP_PV
would cause rows to be removed in Python that still existed in SAS.
This threw off the calculations. Changed early code to change these
missing values to 0's and the missing rows are readded and the
calculations run correctly. Currently get a type mismatch error in the
final dataframe equal assertion in the test, where SAS uses a float of
1.0000 (for example) and Pandas uses an integer of 1.


Former-commit-id: 7c29bbffa9c859e8016bce5f74f048e21c39d712