Add rounding to shift results closer to SAS
Rounding has been added to various columns in the final tested datasets in order to bring the values as close to the SAS results as possible. Have also converted some columns to float columns from integers to fix a type error in the test. Test WILL still fail currently as the two result sets are not identical but I have confirmed that the differences are only due to the use of Python3 rounding. Therefore, the SAS result set will eventually be replaced with a new standard result set which will allow the test to pass.


Former-commit-id: 412f7eb26cf9cdf989e27d90977cebd73694a2ad