Fares unit test written and tested. All passing.
Change made to fares function where INTDATE was being used, this value was coming is as a numeric which dropped the leading '0's of the variables in the columns. This issue is resolved and the data matches aside from the expected rounding difference. The expected result data has now been changed to match the values resulting from the bankers rounding technique implemented as default in Python 3.


Former-commit-id: 189ba5ed59f690b9b35bed923a999a0152d9dd40